# packages/contracts/auth/openapi.yaml
openapi: 3.1.0
info:
  title: Auth API
  version: "1.0.0"
  description: >
    Autenticación y gestión de usuarios Admin/Staff. Los pacientes NO tienen
    cuenta ni login (ver RFC-001-bounded-contexts.md, decisión 1) — este
    servicio solo cubre identidades que acceden al panel administrativo.
    Publica eventos de dominio que otros servicios consumen para mantener
    sus propios read-models (ver ADR-002).
servers:
  - url: /v1
    description: Versión actual. Ver ADR-003-versionado-apis.md para política de versionado.
paths:
...
  /auth/refresh:
    post:
      summary: Renovar accessToken usando un refreshToken válido
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: Nuevo accessToken emitido
          content:
            application/json:
              schema:
                type: object
                required: [accessToken, expiresIn]
                properties:
                  accessToken: { type: string }
                  expiresIn: { type: integer }
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string, example: "UNAUTHORIZED" }
            message: { type: string }
            requestId: { type: string, format: uuid }

    # --- Esquemas de eventos publicados por Auth (Redis Streams) ---
    # No son endpoints HTTP. Se documentan aquí porque son parte del contrato
    # público del servicio (ver ADR-001 y ADR-002). Versionan independiente
    # de las rutas HTTP, ver ADR-003.
...
  responses:
    Unauthorized:
      description: Credenciales inválidas o token inválido/expirado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...